requirements/base.txt
Django==4.2.4
django-crontab==0.7.1
django-import-export==3.3.1
djangorestframework==3.14.0
jdatetime==4.1.1
Pillow==10.0.0
python-dateutil==2.8.2
python-telegram-bot==20.4
pytz==2023.3
requests==2.31.0